+++ CHANGELOG.md
@@ -659,3 +659,5 @@
 * Define `yi_yi_yoklKL` operator
 * Define `same` expression
 * Define `match` expression
+* Define `Cascading'` type alias
+* Define `Cascading'` pattern synonym